refactor(#3928): split the delta-scan emit path out of block_emit.rs (campsite rule)
The rebase made this a ratchet requirement rather than a preference. Pre-rebase,
block_emit.rs was 760 lines against a 744-line base — UNDER the 800 threshold, so
#3928's 16-line refuse-don't-swallow addition at the partition-header arm was not
a ratchet subject at all. #3721 (PR #3814) took that file to 801 on `main`, so the
same addition now grows an OVER-threshold file: 801 -> 817, which the file-size
component FAILs (measured, this branch's lite round 1).

Split rather than CQLITE_ALLOW_FILE_GROWTH, and split at a responsibility rather
than at a line count: `parse_block_emit_delta` is the one `delta-scan`-gated emit
shape in the file (static rows emitted SEPARATELY, a five-value closure), so it
moves VERBATIM to block_emit/delta_emit.rs with the feature gate hoisted to the
`mod`. No behaviour, signature or visibility change.

block_emit.rs 817 -> 405 (now under threshold); delta_emit.rs 437.
Compile-checked both ways: `--features delta-scan` and default.
